AmareluxoCore/requirements.txt
langchain==0.3.27
langchain_core==0.3.76
langchain_openai==0.3.33
langgraph==0.6.7
mcp_use==1.3.10
python-dotenv==1.1.1
streamlit==1.49.1
